requirements.txt
future==0.15.2
Jinja2==2.8
Markdown==2.6.6
MarkupSafe==0.23
PyYAML==3.12
